backend/requirements.txt
flask
flask-cors
gunicorn
requests
python-aiml
